intranet not isolated version quality module migrating goals return the session skeleton blog csv allowed call transfers

+++ INSIGHTSAPI/excels_processing/views.py
@@ -1,16 +1,17 @@
 """This module contains the views for the excels_processing app."""
 import logging
 import os
+import shutil
+import re
+from datetime import datetime
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from rest_framework.decorators import permission_classes
 from rest_framework.permissions import IsAuthenticated
-import shutil
-import re
 import os
-from datetime import datetime
 import mysql.connector
 from .excel_functions import upload_df_to_table, file_to_data_frame
+from django.db import transaction
 
 
 logger = logging.getLogger("requests")
@@ -44,20 +45,30 @@ def robinson_list(request):
     connection = None
     try:
         connection = mysql.connector.connect(**db_config)
+        cursor = connection.cursor()
         rows = upload_df_to_table(filtered_df, connection, "blacklist", columns_mapping)
-        total_rows = len(data_f)
+        query = "SELECT COUNT(*) AS row_count FROM blacklist;"
+        cursor.execute(query)
+        total_rows = cursor.fetchone()[0]
+        uploaded_rows = len(data_f)
         if rows > 0:
             return Response(
                 {
                     "message": "File processed successfully.",
-                    "total_rows": total_rows,
-                    "rows_updated": rows,
+                    "uploaded_rows": uploaded_rows,
+                    "database_rows": total_rows,
+                    "updated_rows": rows,
                 },
                 status=201,
             )
         else:
             return Response(
-                {"message": "No data was inserted.", "total_rows": total_rows},
+                {
+                    "message": "No data was inserted.",
+                    "uploaded_rows": uploaded_rows,
+                    "database_rows": total_rows,
+                    "rows_updated": rows,
+                },
                 status=200,
             )
     except Exception as error:
@@ -80,12 +91,14 @@ def call_transfer_list(request):
     campaign = str(request.POST.get("campaign")).lower()
     paths = {
         "falabella_old": "/var/servers/falabella/BOGOTA/LLAMADAS_PREDICTIVO/",
-        "falabella_new": "/var/servers/calidad/",
+        "falabella_new": "/var/servers/calidad/Llamadas Banco Falabella/",
+        "test_old": "/var/servers/test/",
+        "test_new": "/var/servers/test/",
     }
     # check if the folder exists
     file = request.FILES["file"]
     data_f = file_to_data_frame(file)
-    required_columns = ["FECHA", "CELULAR"]
+    required_columns = ["FECHA", "NUMERO"]
     # Check if the file has the required columns, and say which ones are missing
     missing_columns = [
         column for column in required_columns if column not in data_f.columns
@@ -94,24 +107,35 @@ def call_transfer_list(request):
         return Response(f"Missing columns: {', '.join(missing_columns)}", status=400)
     # Transfer the calls files to the new path
     pattern = re.compile(r"_(\d+)-")
+    fails = []
+
     for row in data_f.itertuples(index=False):
-        date = datetime.strptime(str(row.FECHA.date()), "%Y-%m-%d")
-        number = str(row.CELULAR)
+        date = datetime.strptime(str(row.FECHA).split(" ", maxsplit=1)[0], "%d/%m/%Y")
+        number = str(row.NUMERO)
         match = None
+        if not os.path.exists(
+            os.path.join(paths[f"{campaign}_old"], date.strftime("%Y/%m/%d/OUT/"))
+        ):
+            fails.append(number)
+            break
         for entry in os.scandir(
-            paths[f"{campaign}_old"] + f"{date.year}/{date.month}/{date.day}/OUT/"
+            os.path.join(paths[f"{campaign}_old"], date.strftime("%Y/%m/%d/OUT/"))
         ):
             if entry.is_file() and entry.name.endswith(".mp3"):
                 # Use the compiled pattern to match the middle number
                 match = pattern.search(entry.name)
                 if match and match.group(1) == number:
-                    # Transfer the file
-                    shutil.copy2(
-                        entry.path,
-                        paths[f"{campaign}_new"] + entry.name,
-                    )
-                    match = True
-                    break
-        if match is True:
-            logger.warning(f"No file found for {number} on {date}")
-    return Response("Files transferred successfully", status=200)
+                    try:
+                        # Transfer the file
+                        shutil.copy2(
+                            entry.path,
+                            os.path.join(paths[f"{campaign}_new"], entry.name),
+                        )
+                        match = True
+                        break
+                    except Exception as error:
+                        logger.critical(error)
+                        return Response(str(error), status=500)
+        if match is not True:
+            fails.append(number)
+    return Response({"message": "Files transferred successfully.", "fails": fails})
